requirements.txt
django==3.1.1
djangorestframework==3.11.1
inflect==4.1.0

# Code quality
black==20.8b1
flake8==3.8.3
flake8-black==0.2.1
flake8-bugbear==20.1.4
flake8-docstrings==1.5.0
flake8-isort==4.0.0
isort==5.4.2
mypy==0.782
pytest==6.0.1

# Dev tools
flit==2.3.0
